restruction of i18n directory
