+++ src/common/i18n/languages.ts
