+++ src/common/i18n/locales/zh_tw.ts
@@ -1,0 +1,397 @@
+import { Texts } from "../text_template";
+
+export const zh_tw: Texts = {
+  electronShogi: "Electron將棋",
+  clear: "清除",
+  open: "開啟",
+  saveOverwrite: "覆蓋檔案",
+  newRecord: "新棋譜",
+  newRecordWithBrackets: "（新棋譜）",
+  openRecord: "打開棋譜",
+  saveRecord: "保存棋譜",
+  saveRecordAs: "另存棋譜",
+  openAutoSavingDirectory: "打開自動保存目錄",
+  exportPositionImage: "輸出局面圖",
+  positionImage: "局面圖",
+  close: "關閉",
+  quit: "離開",
+  editing: "編輯",
+  copyAsKIF: "複製KIF棋譜",
+  copyAsCSA: "複製CSA棋譜",
+  copyAsUSI: "複製USI棋譜",
+  copyAsSFEN: "複製SFEN局面",
+  paste: "貼上",
+  copyRecord: "複製棋譜",
+  asKIF: "KIF形式",
+  asCSA: "CSA形式",
+  asUSIUntilCurrentMove: "USI形式(到目前手數為止)",
+  asUSIAll: "USI形式(全部)",
+  copyPositionAsSFEN: "複製局面(SFEN形式)",
+  pasteRecordOrPosition: "貼上棋譜、局面",
+  appendSpecialMove: "特殊手",
+  deleteMoves: "刪除現在位置後的棋譜",
+  view: "表示",
+  toggleFullScreen: "切換全螢幕",
+  defaultFontSize: "預設字體尺寸",
+  largerFontSize: "增加字體尺寸",
+  smallerFontSize: "縮小字體尺寸",
+  settings: "設定",
+  config: "設定",
+  debug: "偵錯",
+  toggleDevTools: "打開/關閉開發者工具顯示",
+  openAppDirectory: "顯示本程式所在資料夾",
+  openSettingDirectory: "開啟設定檔案所在資料夾",
+  openLogDirectory: "開啟紀錄檔(log)所在資料夾",
+  help: "協助",
+  openWebSite: "官方網站",
+  howToUse: "使用教學",
+  checkForUpdates: "檢查最新版本",
+  game: "對局",
+  player: "玩家",
+  selectFromHistory: "從紀錄選取",
+  noHistory: "目前沒有紀錄",
+  saveHistory: "保存紀錄",
+  version: "版本",
+  gameProgress: "對局過程",
+  allGamesCompleted: "連續對局結束",
+  gameEnded: "對局結束",
+  offlineGame: "本地（單機）對局",
+  csaOnlineGame: "通訊對局（CSA）",
+  csaProtocolOnlineGame: "通訊對局（CSA協定）",
+  csaProtocolV121: "CSA協定 1.2.1 標準",
+  csaProtocolV121WithPVComment: "CSA協定 1.2.1 +思考註解",
+  hostToConnect: "連接之伺服器",
+  portNumber: "連接埠號碼",
+  password: "密碼",
+  showPassword: "顯示密碼",
+  logout: "登出",
+  displayGameResults: "確認戰績",
+  interrupt: "中斷",
+  stopGame: "對局中斷",
+  resign: "投了",
+  draw: "平手",
+  impass: "持将棋",
+  repetitionDraw: "千日手",
+  mate: "詰",
+  mateSearch: "詰搜尋",
+  stopMateSearch: "結束詰搜尋",
+  noMateFound: "在目前的盤面中找不到詰。",
+  timeout: "時間耗盡",
+  foulWin: "反則勝利",
+  foulLose: "反則敗北",
+  enteringOfKing: "入玉勝利",
+  winByDefault: "不戰勝",
+  loseByDefault: "不戰敗",
+  winByDeclaration: "宣言勝利",
+  declareWinning: "勝利宣言",
+  research: "檢討",
+  startResearch: "檢討開始",
+  endResearch: "結束檢討",
+  recordAnalysis: "棋譜解析",
+  analysis: "解析",
+  analyze: "解析開始",
+  stopAnalysis: "中斷解析",
+  setupPosition: "編輯局面",
+  startPositionSetup: "開始編輯局面",
+  completePositionSetup: "結束編輯局面",
+  changeTurn: "變更手番",
+  initializePosition: "初始化局面",
+  appSettings: "程式設定",
+  language: "語言",
+  theme: "主題",
+  standard: "標準",
+  green: "綠色",
+  cherryBlossom: "櫻花",
+  customImage: "自定義圖片",
+  autumn: "紅葉",
+  snow: "雪",
+  dark: "深色主題",
+  piece: "棋駒",
+  singleKanjiPiece: "一文字駒",
+  singleKanjiGothicPiece: "一文字駒（黑體）",
+  singleKanjiDarkPiece: "一文字駒（深色）",
+  singleKanjiGothicDarkPiece: "一文字駒（黑體・深色）",
+  backgroundImage: "背景圖片",
+  board: "棋盤",
+  pieceStand: "駒台",
+  lightWoodyTexture: "木目（亮色）",
+  warmWoodTexture: "木目（暖色）",
+  regin: "樹脂",
+  displayFileAndRank: "顯示段・筋",
+  displayLeftControls: "顯示左側操作按鈕",
+  displayRightControls: "顯示右側操作按鈕",
+  tabViewStyle: "分頁顯示形式",
+  oneColumn: "1列",
+  twoColumns: "2列",
+  sounds: "音效",
+  pieceLoudness: "棋駒音效",
+  clockLoudness: "棋鐘音效大小",
+  clockPitch: "棋鐘音效頻率",
+  clockSoundTarget: "棋鐘音效對象",
+  anyTurn: "所有手番",
+  onlyHumanTurn: "只有玩家手番",
+  textEncoding: "文字編碼",
+  strict: "檔案原始編碼",
+  autoDetect: "自動判定",
+  newlineCharacter: "換行符號",
+  old90sMac: "90年代Mac",
+  autoSavingDirectory: "棋譜自動保存地點",
+  select: "選擇",
+  usiProtocol: "USI協定",
+  translateOptionName: "選項名稱翻譯",
+  functionalOnJapaneseOnly: "只有在日文選擇時有效",
+  maxStartupTime: "最大起動等待時間",
+  forDevelopers: "開發者用",
+  enableAppLog: "輸出程式 log",
+  enableUSILog: "輸出USI通信 log",
+  enableCSALog: "輸出CSA通信 log",
+  logLevel: "log 等級",
+  engineSettings: "引擎設定",
+  flipBoard: "盤面反轉",
+  file: "檔案",
+  recordFile: "棋譜檔案",
+  executableFile: "可執行檔案",
+  imageFile: "圖片檔案",
+  remove: "刪除",
+  deleteMove: "刪除該手",
+  recordProperties: "棋譜情報",
+  comments: "備註",
+  moveComments: "棋步備註",
+  searchLog: "思考",
+  pv: "読み筋",
+  mateShort: "詰死",
+  displayPVShort: "再現",
+  evaluation: "評價值",
+  eval: "評價値",
+  estimatedWinRate: "期待勝率",
+  evaluationAndEstimatedWinRate: "評價値・期待勝率",
+  swapEachTurnChange: "手番側有利時為正值",
+  alwaysSenteIsPositive: "先手有利時為正值",
+  signOfEvaluation: "評價值符號",
+  winRateCoefficient: "勝率換算係數",
+  hideTabView: "最小化",
+  expandTabView: "展開分頁",
+  sente: "先手",
+  senteOrShitate: "先手（下手）",
+  gote: "後手",
+  goteOrUwate: "後手（上手）",
+  swapSenteGote: "先後交換",
+  currentPosition: "現在局面",
+  time: "時間",
+  enableEngineTimeout: "開啟引擎時間限制",
+  others: "其他",
+  nextTurn: "手番",
+  elapsedTime: "消費時間",
+  elapsed: "經過時間",
+  rank: "順位",
+  depth: "深度",
+  searchEngine: "引擎",
+  ponder: "對方手番時運算(Ponder)",
+  numberOfThreads: "執行緒數",
+  multiPV: "多重PV",
+  startPosition: "開始局面",
+  maxMoves: "最大手數",
+  gameRepetition: "連續對局",
+  autoRelogin: "自動重新登入",
+  swapTurnWhenGameRepetition: "每局交換手番",
+  outputComments: "輸出備註",
+  saveRecordAutomatically: "自動保存棋譜",
+  adjustBoardToHumanPlayer: "調整到玩家所在方向",
+  adjustBoardAutomatically: "自動調整盤面方向",
+  startGame: "對局開始",
+  cancelGame: "取消對局",
+  allottedTime: "持時間",
+  byoyomi: "讀秒",
+  increments: "增秒",
+  startEndCriteria: "開始・結束條件",
+  endCriteria1Move: "局面結束條件",
+  outputSettings: "輸出設定",
+  noOutputs: "不輸出",
+  insertCommentToTop: "加入到前方",
+  appendCommentToBottom: "在後方新增",
+  overwrite: "覆寫原檔案",
+  fromPrefix: "從",
+  fromSuffix: "",
+  toPrefix: "到",
+  toSuffix: "",
+  plyPrefix: "第",
+  plySuffix: "手",
+  hoursSuffix: "時間",
+  minutesSuffix: "分",
+  secondsSuffix: "秒",
+  engineManagement: "引擎管理",
+  engineName: "引擎名稱",
+  author: "作者",
+  enginePath: "場所",
+  openDirectory: "開啟資料夾",
+  displayName: "表示名稱",
+  invoke: "執行",
+  resetToEngineDefaultValues: "回復至引擎預設設定",
+  defaultValue: "預設値",
+  noEngineRegistered: "尚未登錄引擎。",
+  duplicate: "複製",
+  add: "追加",
+  recommended: "推薦",
+  import: "匯入",
+  saveAndClose: "保存並關閉",
+  save: "保存",
+  saveAs: "另存為",
+  cancel: "取消",
+  back: "返回",
+  name: "名稱",
+  prediction: "預測",
+  best: "最善",
+  nodes: "Node數",
+  hashUsage: "Hash使用率",
+  nonHandicap: "平手",
+  lanceHandicap: "香落",
+  rightLanceHandicap: "右香落",
+  bishopHandicap: "角落",
+  rookHandicap: "飛車落",
+  rookLanceHandicap: "飛車香落",
+  twoPiecesHandicap: "二枚落",
+  fourPiecesHandicap: "四枚落",
+  sixPiecesHandicap: "六枚落",
+  eightPiecesHandicap: "八枚落",
+  tsumeShogi: "詰將棋",
+  doubleKingTsumeShogi: "雙玉詰將棋",
+  startDateTime: "開始日時",
+  endDateTime: "結束日時",
+  gameDate: "對局日",
+  tournament: "棋戰",
+  strategy: "戰型",
+  gameTitle: "標題",
+  timeLimit: "持時間",
+  place: "場所",
+  postedOn: "登錄於",
+  note: "備註",
+  senteShortName: "先手省略名",
+  goteShortName: "後手省略名",
+  opusNo: "作品編號",
+  opusName: "作品名",
+  publishedBy: "發表於",
+  publishedOn: "發表年月",
+  source: "來源",
+  numberOfMoves: "手數",
+  integrity: "完全性",
+  recordCategory: "分類",
+  award: "受賞",
+  filterByOptionName: "搜尋設定名稱",
+  filterByEngineName: "搜尋引擎名稱",
+  bookStyle: "書籍風",
+  gameStyle: "對局畫面風",
+  none: "無",
+  bgCover: "圖片置中",
+  bgContain: "擴大表示",
+  bgTile: "磁磚狀表示",
+  inaccuracy: "緩手",
+  dubious: "疑問手",
+  mistake: "惡手",
+  blunder: "大惡手",
+  inaccuracyThreshold: "緩手閾値",
+  dubiousThreshold: "疑問手閾値",
+  mistakeThreshold: "惡手閾値",
+  blunderThreshold: "大惡手閾値",
+  typeCustomTitleHere: "輸入自定義標題",
+  displayEmptyElements: "顯示未定義資料",
+  waitingForNewGame: "正在等待下一場對局開始。",
+  tryingToConnectAndLoginToCSAServer: "正在連接並登入CSA伺服器。",
+  inBrowserLogsOutputToConsoleAndIgnoreThisSetting:
+    "※在瀏覽器版本中 log 會於 console 中顯示，並無視此處的設定。",
+  shouldRestartToApplyLogSettings:
+    "※您需要重新啟動本程式以使用變更後的 log 設定。",
+  canOpenLogDirectoryFromMenu:
+    "※log 的輸出檔案可以在「偵錯」-「開啟紀錄檔案資料夾」開啟。",
+  hasNoOldLogCleanUpFeature: "※現在並沒有舊 log 的自動刪除機制。",
+  processingPleaseWait: "現在處理中。請稍待一會。",
+  importingFollowingRecordOrPosition: "將匯入以下棋譜（局面）。",
+  supportsKIFCSAUSI: "※支援KIF/CSA/SFEN形式。",
+  plesePasteRecordIntoTextArea: "※請在文字輸入區域貼上您的棋譜。",
+  desktopVersionPastesAutomatically: "※安裝程式版將會自動貼上棋譜。",
+  someLogsDisabled: "部份 log 已被無效化。",
+  logsRecommendedForCSAProtocol: "若使用CSA協定對局，建議輸出各項 log 。",
+  pleaseEnableLogsAndRestart: "請在程式設定中開啟 log 並重新啟動本程式。",
+  notSendPVOnStandardCSAProtocol:
+    "在標準的CSA協定中不會送出評價值以及思考棋步。",
+  sendPVDoNotUseOnWCSC:
+    "使用Floodgate形式傳送評價值與思考棋步。請不要在WCSC中使用。",
+  csaProtocolSendPlaintextPassword: "在CSA協定中，密碼為明文傳輸。",
+  passwordWillSavedPlaintextBecauseOSSideEncryptionNotAvailable:
+    "由於無法使用系統的加密機能，輸入的密碼將會以明文保存。",
+  pleaseUncheckSaveHistoryIfNotWantSave:
+    "若不想保存密碼，請不要將「保存紀錄」勾選。",
+  csaProtocolSendPlaintextPasswordRegardlessOfHistory:
+    "不過，CSA協定仍會以明文傳輸您的密碼。",
+  areYouSureWantToQuitGames: "要中斷連續對局嗎？",
+  areYouSureWantToRequestQuit: "若提出中斷要求，可能會被判負。請問您要繼續嗎？",
+  areYouSureWantToClearRecord: "將會刪除現在的棋譜。請問您要繼續嗎？",
+  areYouSureWantToDiscardPosition: "將不會保存現在的局面。請問您要繼續嗎？",
+  youCanNotCloseAppWhileCSAOnlineGame:
+    "由於CSA協定正在使用中，本程式無法被關閉。",
+  fileExtensionNotSupported: "無法支援的副檔名。",
+  errorOccuredWhileDisconnectingFromCSAServer:
+    "在與CSA伺服器中斷連線時發生錯誤。",
+  failedToConnectToCSAServer: "無法連接CSA伺服器。",
+  disconnectedFromCSAServer: "與CSA伺服器的連接結束。",
+  csaServerLoginDenied: "您對CSA伺服器的登入被拒絕。",
+  thisFeatureNotAvailableOnWebApp: "Web版無法使用本機能",
+  failedToSendGoCommand: "無法送出go指令。",
+  failedToSendPonderCommand: "無法送出ponder指令。",
+  failedToSendStopCommand: "無法送出stop指令。",
+  failedToSaveRecord: "棋譜保存失敗。",
+  failedToParseSFEN: "SFEN讀取失敗。",
+  failedToDetectRecordFormat: "無法判別棋譜形式。",
+  unknownFileExtension: "未知的檔案形式。",
+  emptyRecordInput: "您尚未輸入棋譜。",
+  invalidPieceName: "不合法的棋駒",
+  invalidTurn: "不合法的手番",
+  invalidMove: "不合法的棋步",
+  invalidMoveNumber: "不合法的手數",
+  invalidDestination: "不合法的移動目的地",
+  pieceNotExists: "不存在的棋駒",
+  invalidLine: "不存在的行",
+  invalidHandicap: "不合法的手合配置",
+  invalidBoard: "不合法的盤面",
+  invalidHandPiece: "不合法的持駒",
+  invalidUSI: "不合法的USI",
+  backgroundImageFileNotSelected: "尚未選取背景圖片。",
+  boardImageFileNotSelected: "尚未選取盤面圖片。",
+  pieceStandImageFileNotSelected: "尚未選取駒台圖片。",
+  pieceVolumeMustBe0To100Percent: "請在0%～100%之間指定棋駒音效大小。",
+  clockVolumeMustBe0To100Percent: "請在0%～100%之間指定棋鐘音效大小。",
+  clockPitchMustBe220To880Hz: "請在220Hz～880Hz之間指定棋鐘音效頻率。",
+  engineTimeoutMustBe1To300Seconds: "請在1秒～300秒之間指定引擎執行最長時間。",
+  coefficientInSigmoidMustBeGreaterThan0: "請將勝率換算係數填為大於0之值。",
+  inaccuracyThresholdMustBe1To100Percent: "請在0%～100%之間指定緩手門檻。",
+  dubiousThresholdMustBe1To100Percent: "請在0%～100%之間指定疑問手門檻。",
+  mistakeThresholdMustBe1To100Percent: "請在0%～100%之間指定惡手門檻。",
+  blunderThresholdMustBe1To100Percent: "請在0%～100%之間指定大惡手門檻。",
+  inaccuracyThresholdMustBeLessThanDubiousThreshold:
+    "緩手門檻應小於疑問手門檻。",
+  dubiousThresholdMustBeLessThanMistakeThreshold: "疑問手門檻應小於惡手門檻。",
+  mistakeThresholdMustBeLessThanBlunderThreshold: "惡手門檻應小於大惡手門檻。",
+  thisEngineNotSupportsMateSearch: "這個引擎不支援詰將棋搜索。",
+  tryToReloginToCSAServerNSecondsLater: (n) =>
+    `請在${n}秒後再次嘗試登入 CSA 伺服器。`,
+  mateInNPlyDoYouWantToDisplay: (n) => `尋找到${n}手詰。要顯示結果嗎？`,
+  errorsOccurred: (n) => `發生 ${n} 種類的錯誤。`,
+  between: (a, b) => `自 ${a} 到 ${b} `,
+  addNthEngine: (n) => `追加第 ${n} 個引擎`,
+  copyOf: (name) => `${name} 的拷貝`,
+  keepLatest: (n) => `到最新${n}件`,
+  areYouSureWantToDeleteFollowingMove: (n) =>
+    `將會刪除${n}手目以後的棋譜。請問您要繼續嗎？`,
+  failedToOpenDirectory: (path) => `無法開啟檔案目錄：${path}`,
+  unexpectedEventSenderPleaseReport(sender) {
+    return `無法預期的事件發送方已發生。請將該錯誤訊息告知開發者，謝謝。 [${sender}]`;
+  },
+  unexpectedHTTPMethodPleaseReport(method) {
+    return `無法預期的 HTTP 方法。請將該錯誤訊息告知開發者，謝謝。 [${method}]`;
+  },
+  unexpectedRequestURLPleaseReport(url) {
+    return `無法從 URL 獲取資訊。請將該錯誤訊息告知開發者，謝謝。 [${url}]`;
+  },
+  noResponseFromEnginePleaseExtendTimeout(seconds) {
+    return `引擎在${seconds}秒内沒有回應。若引擎的啟動時間稍長，請在設定中調整引擎最長等待時間。`;
+  },
+};
