+++ src/common/i18n/text_template.ts
@@ -1,0 +1,374 @@
+export type Texts = {
+  electronShogi: string;
+  clear: string;
+  open: string;
+  saveOverwrite: string;
+  newRecord: string;
+  newRecordWithBrackets: string;
+  openRecord: string;
+  saveRecord: string;
+  saveRecordAs: string;
+  openAutoSavingDirectory: string;
+  exportPositionImage: string;
+  positionImage: string;
+  close: string;
+  quit: string;
+  editing: string;
+  copyAsKIF: string;
+  copyAsCSA: string;
+  copyAsUSI: string;
+  copyAsSFEN: string;
+  paste: string;
+  copyRecord: string;
+  asKIF: string;
+  asCSA: string;
+  asUSIUntilCurrentMove: string;
+  asUSIAll: string;
+  copyPositionAsSFEN: string;
+  pasteRecordOrPosition: string;
+  appendSpecialMove: string;
+  deleteMoves: string;
+  view: string;
+  toggleFullScreen: string;
+  defaultFontSize: string;
+  largerFontSize: string;
+  smallerFontSize: string;
+  settings: string;
+  config: string;
+  debug: string;
+  toggleDevTools: string;
+  openAppDirectory: string;
+  openSettingDirectory: string;
+  openLogDirectory: string;
+  help: string;
+  openWebSite: string;
+  howToUse: string;
+  checkForUpdates: string;
+  game: string;
+  player: string;
+  selectFromHistory: string;
+  noHistory: string;
+  saveHistory: string;
+  version: string;
+  gameProgress: string;
+  allGamesCompleted: string;
+  gameEnded: string;
+  offlineGame: string;
+  csaOnlineGame: string;
+  csaProtocolOnlineGame: string;
+  csaProtocolV121: string;
+  csaProtocolV121WithPVComment: string;
+  hostToConnect: string;
+  portNumber: string;
+  password: string;
+  showPassword: string;
+  logout: string;
+  displayGameResults: string;
+  interrupt: string;
+  stopGame: string;
+  resign: string;
+  draw: string;
+  impass: string;
+  repetitionDraw: string;
+  mate: string;
+  mateSearch: string;
+  stopMateSearch: string;
+  noMateFound: string;
+  timeout: string;
+  foulWin: string;
+  foulLose: string;
+  enteringOfKing: string;
+  winByDefault: string;
+  loseByDefault: string;
+  winByDeclaration: string;
+  declareWinning: string;
+  research: string;
+  startResearch: string;
+  endResearch: string;
+  recordAnalysis: string;
+  analysis: string;
+  analyze: string;
+  stopAnalysis: string;
+  setupPosition: string;
+  startPositionSetup: string;
+  completePositionSetup: string;
+  changeTurn: string;
+  initializePosition: string;
+  appSettings: string;
+  language: string;
+  theme: string;
+  standard: string;
+  green: string;
+  cherryBlossom: string;
+  customImage: string;
+  autumn: string;
+  snow: string;
+  dark: string;
+  piece: string;
+  singleKanjiPiece: string;
+  singleKanjiGothicPiece: string;
+  singleKanjiDarkPiece: string;
+  singleKanjiGothicDarkPiece: string;
+  backgroundImage: string;
+  board: string;
+  pieceStand: string;
+  lightWoodyTexture: string;
+  warmWoodTexture: string;
+  regin: string;
+  displayFileAndRank: string;
+  displayLeftControls: string;
+  displayRightControls: string;
+  tabViewStyle: string;
+  oneColumn: string;
+  twoColumns: string;
+  sounds: string;
+  pieceLoudness: string;
+  clockLoudness: string;
+  clockPitch: string;
+  clockSoundTarget: string;
+  anyTurn: string;
+  onlyHumanTurn: string;
+  textEncoding: string;
+  strict: string;
+  autoDetect: string;
+  newlineCharacter: string;
+  old90sMac: string;
+  autoSavingDirectory: string;
+  select: string;
+  usiProtocol: string;
+  translateOptionName: string;
+  functionalOnJapaneseOnly: string;
+  maxStartupTime: string;
+  forDevelopers: string;
+  enableAppLog: string;
+  enableUSILog: string;
+  enableCSALog: string;
+  logLevel: string;
+  engineSettings: string;
+  flipBoard: string;
+  file: string;
+  recordFile: string;
+  executableFile: string;
+  imageFile: string;
+  remove: string;
+  deleteMove: string;
+  recordProperties: string;
+  comments: string;
+  moveComments: string;
+  searchLog: string;
+  pv: string;
+  mateShort: string;
+  displayPVShort: string;
+  evaluation: string;
+  eval: string;
+  estimatedWinRate: string;
+  evaluationAndEstimatedWinRate: string;
+  swapEachTurnChange: string;
+  alwaysSenteIsPositive: string;
+  signOfEvaluation: string;
+  winRateCoefficient: string;
+  hideTabView: string;
+  expandTabView: string;
+  sente: string;
+  senteOrShitate: string;
+  gote: string;
+  goteOrUwate: string;
+  swapSenteGote: string;
+  currentPosition: string;
+  time: string;
+  enableEngineTimeout: string;
+  others: string;
+  nextTurn: string;
+  elapsedTime: string;
+  elapsed: string;
+  rank: string;
+  depth: string;
+  searchEngine: string;
+  ponder: string;
+  numberOfThreads: string;
+  multiPV: string;
+  startPosition: string;
+  maxMoves: string;
+  gameRepetition: string;
+  autoRelogin: string;
+  swapTurnWhenGameRepetition: string;
+  outputComments: string;
+  saveRecordAutomatically: string;
+  adjustBoardToHumanPlayer: string;
+  adjustBoardAutomatically: string;
+  startGame: string;
+  cancelGame: string;
+  allottedTime: string;
+  byoyomi: string;
+  increments: string;
+  startEndCriteria: string;
+  endCriteria1Move: string;
+  outputSettings: string;
+  noOutputs: string;
+  insertCommentToTop: string;
+  appendCommentToBottom: string;
+  overwrite: string;
+  fromPrefix: string;
+  fromSuffix: string;
+  toPrefix: string;
+  toSuffix: string;
+  plyPrefix: string;
+  plySuffix: string;
+  hoursSuffix: string;
+  minutesSuffix: string;
+  secondsSuffix: string;
+  engineManagement: string;
+  engineName: string;
+  author: string;
+  enginePath: string;
+  openDirectory: string;
+  displayName: string;
+  invoke: string;
+  resetToEngineDefaultValues: string;
+  defaultValue: string;
+  noEngineRegistered: string;
+  duplicate: string;
+  add: string;
+  recommended: string;
+  import: string;
+  saveAndClose: string;
+  save: string;
+  saveAs: string;
+  cancel: string;
+  back: string;
+  name: string;
+  prediction: string;
+  best: string;
+  nodes: string;
+  hashUsage: string;
+  nonHandicap: string;
+  lanceHandicap: string;
+  rightLanceHandicap: string;
+  bishopHandicap: string;
+  rookHandicap: string;
+  rookLanceHandicap: string;
+  twoPiecesHandicap: string;
+  fourPiecesHandicap: string;
+  sixPiecesHandicap: string;
+  eightPiecesHandicap: string;
+  tsumeShogi: string;
+  doubleKingTsumeShogi: string;
+  startDateTime: string;
+  endDateTime: string;
+  gameDate: string;
+  tournament: string;
+  strategy: string;
+  gameTitle: string;
+  timeLimit: string;
+  place: string;
+  postedOn: string;
+  note: string;
+  senteShortName: string;
+  goteShortName: string;
+  opusNo: string;
+  opusName: string;
+  publishedBy: string;
+  publishedOn: string;
+  source: string;
+  numberOfMoves: string;
+  integrity: string;
+  recordCategory: string;
+  award: string;
+  filterByOptionName: string;
+  filterByEngineName: string;
+  bookStyle: string;
+  gameStyle: string;
+  none: string;
+  bgCover: string;
+  bgContain: string;
+  bgTile: string;
+  inaccuracy: string;
+  dubious: string;
+  mistake: string;
+  blunder: string;
+  inaccuracyThreshold: string;
+  dubiousThreshold: string;
+  mistakeThreshold: string;
+  blunderThreshold: string;
+  typeCustomTitleHere: string;
+  displayEmptyElements: string;
+  waitingForNewGame: string;
+  tryingToConnectAndLoginToCSAServer: string;
+  inBrowserLogsOutputToConsoleAndIgnoreThisSetting: string;
+  shouldRestartToApplyLogSettings: string;
+  canOpenLogDirectoryFromMenu: string;
+  hasNoOldLogCleanUpFeature: string;
+  processingPleaseWait: string;
+  importingFollowingRecordOrPosition: string;
+  supportsKIFCSAUSI: string;
+  plesePasteRecordIntoTextArea: string;
+  desktopVersionPastesAutomatically: string;
+  someLogsDisabled: string;
+  logsRecommendedForCSAProtocol: string;
+  pleaseEnableLogsAndRestart: string;
+  notSendPVOnStandardCSAProtocol: string;
+  sendPVDoNotUseOnWCSC: string;
+  csaProtocolSendPlaintextPassword: string;
+  passwordWillSavedPlaintextBecauseOSSideEncryptionNotAvailable: string;
+  pleaseUncheckSaveHistoryIfNotWantSave: string;
+  csaProtocolSendPlaintextPasswordRegardlessOfHistory: string;
+  areYouSureWantToQuitGames: string;
+  areYouSureWantToRequestQuit: string;
+  areYouSureWantToClearRecord: string;
+  areYouSureWantToDiscardPosition: string;
+  youCanNotCloseAppWhileCSAOnlineGame: string;
+  fileExtensionNotSupported: string;
+  errorOccuredWhileDisconnectingFromCSAServer: string;
+  failedToConnectToCSAServer: string;
+  disconnectedFromCSAServer: string;
+  csaServerLoginDenied: string;
+  thisFeatureNotAvailableOnWebApp: string;
+  failedToSendGoCommand: string;
+  failedToSendPonderCommand: string;
+  failedToSendStopCommand: string;
+  failedToSaveRecord: string;
+  failedToParseSFEN: string;
+  failedToDetectRecordFormat: string;
+  unknownFileExtension: string;
+  emptyRecordInput: string;
+  invalidPieceName: string;
+  invalidTurn: string;
+  invalidMove: string;
+  invalidMoveNumber: string;
+  invalidDestination: string;
+  pieceNotExists: string;
+  invalidLine: string;
+  invalidHandicap: string;
+  invalidBoard: string;
+  invalidHandPiece: string;
+  invalidUSI: string;
+  backgroundImageFileNotSelected: string;
+  boardImageFileNotSelected: string;
+  pieceStandImageFileNotSelected: string;
+  pieceVolumeMustBe0To100Percent: string;
+  clockVolumeMustBe0To100Percent: string;
+  clockPitchMustBe220To880Hz: string;
+  engineTimeoutMustBe1To300Seconds: string;
+  coefficientInSigmoidMustBeGreaterThan0: string;
+  inaccuracyThresholdMustBe1To100Percent: string;
+  dubiousThresholdMustBe1To100Percent: string;
+  mistakeThresholdMustBe1To100Percent: string;
+  blunderThresholdMustBe1To100Percent: string;
+  inaccuracyThresholdMustBeLessThanDubiousThreshold: string;
+  dubiousThresholdMustBeLessThanMistakeThreshold: string;
+  mistakeThresholdMustBeLessThanBlunderThreshold: string;
+  thisEngineNotSupportsMateSearch: string;
+  tryToReloginToCSAServerNSecondsLater: (n: number) => string;
+  mateInNPlyDoYouWantToDisplay: (n: number) => string;
+  errorsOccurred: (n: number) => string;
+  between: (a: unknown, b: unknown) => string;
+  addNthEngine: (n: number) => string;
+  copyOf: (name: string) => string;
+  keepLatest: (n: number) => string;
+  areYouSureWantToDeleteFollowingMove: (n: number) => string;
+  failedToOpenDirectory: (path: string) => string;
+  unexpectedEventSenderPleaseReport: (sender: string) => string;
+  unexpectedHTTPMethodPleaseReport: (method: string) => string;
+  unexpectedRequestURLPleaseReport: (url: string) => string;
+  noResponseFromEnginePleaseExtendTimeout: (seconds: number) => string;
+};
